Add co_stacksize field to codeobject structure, and stacksize argument to PyCode_New() argument list.  Move MAXBLOCKS constant to conpile.h.
Added accurate calculation of the actual stack size needed by the
generated code.

Also commented out all fprintf statements (except for a new one to
diagnose stack underflow, and one in #ifdef'ed out code), and added
some new TO DO suggestions (now that the stacksize is taken of the TO
DO list).


git-svn-id: http://svn.python.org/projects/python/trunk@7466 6015fed2-1504-0410-9fe1-9d1591cc4771