Add a single Python-wide (!) lock on import.  Only one thread at a time can be in PyImport_ImportModuleEx().  Recursive calls from the same thread are okay.
Potential problems:

- The lock should really be part of the interpreter state rather than
global, but that would require modifying more files, and I first want
to figure out whether this works at all.

- One could argue that the lock should be per module -- however that
would be complicated to implement.  We would have to have a linked
list of locks per module name, *or* invent a new object type to
represent a lock, so we can store the locks in the module or in a
separate dictionary.  Both seem unwarranted.  The one situation where
this can cause problems is when loading a module takes a long time,
e.g. when the module's initialization code interacts with the user --
during that time, no other threads can run.  I say, "too bad."


git-svn-id: http://svn.python.org/projects/python/trunk@10011 6015fed2-1504-0410-9fe1-9d1591cc4771