Hacked for Win32 by Mark Hammond. Reformatted for 8-space tabs and fitted into 80-char lines by GvR.
Mark writes:

* the Win32 version now accepts the same args as the Unix version.
The win32 specific "tag" param is now optional.  The end result is
that the exact same test suite runs on Windows (definately a worthy
goal!).

* I changed the error object.  All occurences of the error, except
for 1, corresponds to an underlying OS error.  This one was changed
to a ValueError (a better error for that condition), and the module
error object is now simply EnvironmentError.  All win32 error
routines now call the new Windows specific error handler.


git-svn-id: http://svn.python.org/projects/python/trunk@14890 6015fed2-1504-0410-9fe1-9d1591cc4771