Eric Raymond:
(1) Added and documented the capability for shlex to handle
lexical-level inclusion and a stack of input sources.  Also, the input
stream member is now documented, and the constructor takes an optional
source-filename.  The class provides facilities to generate error
messages that track file and line number.

(2) Add a convenience function to generate C-compiler style error
leaders.


git-svn-id: http://svn.python.org/projects/python/trunk@15301 6015fed2-1504-0410-9fe1-9d1591cc4771