Generalize operator.indexOf (PySequence_Index) to work with any iterable object.  I'm not sure how that got overlooked before!
Got rid of the internal _PySequence_IterContains, introduced a new
internal _PySequence_IterSearch, and rewrote all the iteration-based
"count of", "index of", and "is the object in it or not?" routines to
just call the new function.  I suppose it's slower this way, but the
code duplication was getting depressing.


git-svn-id: http://svn.python.org/projects/python/trunk@23063 6015fed2-1504-0410-9fe1-9d1591cc4771