SF bug 533234:  tm_isdst > 1 Passed to strftime. One more time on this turkey, but duller instead of cleverer.
Curious:  The docs say __getslice__ has been deprecated since 2.0, but
list.__getitem__ still doesn't work if you pass it a slice.  This makes
it a lot clearer to emulate a list by *being* a list <wink>.

Bugfix candidate.  Michael, just pile this patch on top of the others
that went by -- no need to try to pick these apart.


git-svn-id: http://svn.python.org/projects/python/trunk@25834 6015fed2-1504-0410-9fe1-9d1591cc4771