Implement an idea by Paul Rubin:
Change pickling format for bools to use a backwards compatible
encoding.  This means you can pickle True or False on Python 2.3
and Python 2.2 or before will read it back as 1 or 0.  The code
used for pickling bools before would create pickles that could
not be read in previous Python versions.


git-svn-id: http://svn.python.org/projects/python/trunk@26212 6015fed2-1504-0410-9fe1-9d1591cc4771