Changed import from     from test.test_support import TestSkipped, run_unittest to     from test_support import TestSkipped, run_unittest
Otherwise, if the Japanese codecs aren't installed, regrtest doesn't
believe the TestSkipped exception raised by this test matches the

    except (ImportError, test_support.TestSkipped), msg:

it's looking for, and reports the skip as a crash failure instead of
as a skipped test.

I suppose this will make it harder to run this test outside of
regrtest, but under the assumption only Barry does that, better to
make it skip cleanly for everyone else.


git-svn-id: http://svn.python.org/projects/python/trunk@27727 6015fed2-1504-0410-9fe1-9d1591cc4771