parse(), _parseheaders(), _parsebody(): A fix for SF bug #633527, where in lax parsing, the first non-header line after a header block (e.g. the first line not containing a colon, and not a continuation), can be treated as the first body line, even without the RFC mandated blank line separator.
rfc822 had this behavior, and I vaguely remember problems with this,
but can't remember details.  In any event, all the tests still pass,
so I guess we'll find out. ;/

This patch works by returning the non-header, non-continuation line
from _parseheader() and using that as the first header line prepended
to fp.read() if given.  It's usually None.

We use this approach instead of trying to seek/tell the file-like
object.


git-svn-id: http://svn.python.org/projects/python/trunk@29533 6015fed2-1504-0410-9fe1-9d1591cc4771