Change warnings to avoid importing re module during startup.
Add API function simplefilter() that does not create or install
regular expressions to match message or module.  Extend the filters
data structure to store None as an alternative to re.compile("").

Move the _test() function to test_warnings and add some code to try
and avoid disturbing the global state of the warnings module.


git-svn-id: http://svn.python.org/projects/python/trunk@33487 6015fed2-1504-0410-9fe1-9d1591cc4771