To avoid problems with conflicting dlls, the windows installers built by bdist_wininst *must* use the same runtime libary as the Python version.
Actually this means the Python version where the installer is run, not
the one which is used to build it.  Must think about that - for now I
assume MSVC6 is used up to Python 2.3, and MSVC7.1 is used starting at
Python 2.4.

So the filename for wininst.exe is now wininst-6.exe for the Release
version and wininst-6_d.exe for the Debug version, when built with
MSVC6.


git-svn-id: http://svn.python.org/projects/python/trunk@35327 6015fed2-1504-0410-9fe1-9d1591cc4771