When importing an extension on Windows, the code reads the PE 'import table' of the dll, to make sure that the dll really was build for the correct Python version.  It does this by looking for an entry 'pythonXY.dll' (X.Y is the Python version number).
The code now checks the size of the dll's import table before reading
entries from it.  Before this patch, the code crashed trying to read
the import table when the size was zero (as in Win2k's wmi.dll, for
example).

Look for imports of 'pythonXY_d.dll' in a debug build instead of
'pythonXY.dll'.

Fixes SF 951851: Crash when reading "import table" of certain windows dlls.

Already backported to the 2.3 branch.


git-svn-id: http://svn.python.org/projects/python/trunk@36289 6015fed2-1504-0410-9fe1-9d1591cc4771