_do_a_fancy_diff():  Pay no attention to the ellipses behind the curtain.
While a fancy diff can be confusing in the presence of ellipses, so far
I'm finding (2-0-0) that it's much more a major aid in narrowing down the
possibilities when an ellipsis-slinging test fails.  So we no longer
refuse to do a fancy diff just because of ellipses.

This isn't ideal; it's just better.


git-svn-id: http://svn.python.org/projects/python/trunk@37157 6015fed2-1504-0410-9fe1-9d1591cc4771