The hppa ubuntu box sometimes hangs forever in these tests.  My guess is that the wait is failing for some reason.  Use WNOHANG, so we won't wait until the buildbot kills the test suite.
I haven't been able to reproduce the failure, so I'm not sure if
this will help or not.  Hopefully, this change will cause the test
to fail, rather than hang.  That will be better since we will get
the rest of the test results.  It may also help us debug the real problem.


git-svn-id: http://svn.python.org/projects/python/trunk@47014 6015fed2-1504-0410-9fe1-9d1591cc4771