Fix for error reported by Michael Brenner via the Stackless mailing list.
When the interpreter exits, all the remaining tasklets are killed via slp_kill_tasks_with_stacks in the main tasklet.  However, it is not possible to loop over the tasklets one by one killing them due to how scheduling works and that you need to schedule a tasklet to kill it.  When a tasklet is scheduled, and it finishes or blocks, the next tasklet in the list is then scheduled.  So what would happen is that every tasklet in the scheduler between the tasklet to be killed and the main tasklet, would be scheduled before we were back to the main tasklet and the killing of tasklets.  This is definitely not what we want.

The fix moves the main tasklet after the one we are killing, so that we are directly back to the tasklet killing loop when the one being killed exits.  At this time there is no reproducibility case to include as a unit test, so the script which reproduces the problem is included in the test directory.


git-svn-id: http://svn.python.org/projects/stackless/Python-2.4.3/dev@51422 6015fed2-1504-0410-9fe1-9d1591cc4771