Merged in r64233-64772 from the mainline trunk.
'configure' is still an older version generated from an older version of 'configure.in'.  It still needs to be regenerated and checked in.

There are still deprecation warnings happening when you subclass the tasklet class and have it call the base '__init__' function with the passed in function as an argument when it gets instantiated.  This appears when running the Stackless unittests, in 'test_watchdog.py'.

The mainline 'test_sys.py' unittest is failing because it checks the size of the typeobject according to a formula, which does not take into account what Stackless adds to this structure.

One mainline unittest fails, 'test_distutils.py', however, this also fails when Stackless is compiled as normal Python with the Stackless functionality turned off - so it is not a Stackless problem.

git-svn-id: http://svn.python.org/projects/stackless/trunk@64776 6015fed2-1504-0410-9fe1-9d1591cc4771