Ported r61053 from release25-maint:
When an installation of Stackless Python is built for Windows, the Python header files included are the Stackless modified ones which depend on some of the Stackless header files.  These were not previously included in the installer build process.

Now they are, and distutils has been modified to be aware of the new Stackless include directory, and to use it when compiling.


git-svn-id: http://svn.python.org/projects/stackless/trunk@64780 6015fed2-1504-0410-9fe1-9d1591cc4771